Auth fix

+++ front/src/components/Profile/index.tsx
@@ -5,7 +5,7 @@ import {
 } from "react"
 import { useNavigate } from "react-router"
 import Cookies from "js-cookie"
-import axios from "axios"
+import axios, { AxiosError } from "axios"
 
 import {
 	Style,
@@ -23,6 +23,7 @@ import { User, UserAuthenticate } from "../../utils/types"
 
 import deconnexionIcon from "../../assets/deconnexion.png"
 import settingsIcon from "../../assets/settings.png"
+import { ErrorResponse, SettingData } from '../../utils/types'
 
 type PropsProfile = {
 	userAuthenticate: UserAuthenticate,
@@ -58,23 +59,33 @@ function Profile({ userAuthenticate, card, displayCard, userTarget, setUserTarge
 	async function handleDeconnexionClickButton() {
 		try {
 
-			await axios.get(`http://${url}:3333/auth/logout`, {
+			const response: boolean = await axios.get(`http://${url}:3333/auth/logout`, {
 				headers: {
 					'Authorization': `Bearer ${token}`
 				}
 			}) 
 			
-			Cookies.remove('access_token')
-			Cookies.remove('id')
-			Cookies.remove('two_FA')
-			//localStorage.removeItem('token')
-			localStorage.clear();
-			setToken('')
-			navigate("/")
+			if (response) {
+				Cookies.remove('access_token')
+				Cookies.remove('id')
+				Cookies.remove('two_FA')
+				Cookies.remove('isNew')
+				localStorage.clear();
+				setToken('')
+				navigate("/")
+			}
 		}
 		catch (error) {
-			console.log(error)
+			if (axios.isAxiosError(error)) {
+				const axiosError = error as AxiosError<ErrorResponse>
+				const { statusCode } = axiosError.response?.data!
+				console.log(error.message)
+				console.log(statusCode)
+			}
+			else
+				navigate("/error");
 		}
+		
 	}
 
 	return (
